Match role_id to the tinyint that list_roles actually uses
list_roles.id is a legacy tinyIncrements() (tinyint unsigned). Declaring
role_id as unsignedInteger made MySQL reject the foreign key outright, and
because DDL is not transactional the table survived the failed migration
without its key or its seeded rows -- blocking every retry.

role_permissions hit this same wall and documents the fix; follow it.

The suite runs on in-memory SQLite, which does not enforce foreign key type
matching, so no test could have caught this.

+++ database/migrations/2026_09_14_000006_create_supervisor_action_roles_table.php
@@ -23,7 +23,9 @@ return new class extends Migration
         Schema::create('supervisor_action_roles', function (Blueprint $table) {
             $table->increments('id');
             $table->string('action', 60);
-            $table->unsignedInteger('role_id');
+            // list_roles.id is a legacy tinyIncrements() (tinyint unsigned); a foreign
+            // key across mismatched integer widths is rejected outright by MySQL.
+            $table->unsignedTinyInteger('role_id');
             $table->timestamps();
 
             $table->unique(['action', 'role_id']);
